Isolated tests for long functions (#8)
These functions (with the exception of `parall.gmj`) are explicitly
mentioned on README.md, and the execution of their examples is not
straightforward, so their test scripts are separated. Perhaps using the
function examples is not the way to go, and new, shorter (sub 5-second) test code needs
to be written before this issue is resolved.